+++ querydsl-jdo/src/main/java/com/querydsl/jdo/sql/AbstractSQLQuery.java
@@ -99,7 +99,7 @@ public abstract class AbstractSQLQuery<T, Q extends AbstractSQLQuery<T, Q>> exte
         reset();
         Long rv = (Long) execute(query, true);
         if (rv != null) {
-            return rv.longValue();
+            return rv;
         } else {
             throw new QueryException("Query returned null");
         }
