+++ querydsl-sql/src/main/java/com/querydsl/sql/types/ByteType.java
@@ -42,7 +42,7 @@ public class ByteType extends AbstractType<Byte> {
     @Override
     public Byte getValue(ResultSet rs, int startIndex) throws SQLException {
         byte val = rs.getByte(startIndex);
-        return rs.wasNull() ? null : Byte.valueOf(val);
+        return rs.wasNull() ? null : val;
     }
 
     @Override
