+++ querydsl-sql/src/main/java/com/querydsl/sql/types/ShortType.java
@@ -42,7 +42,7 @@ public class ShortType extends AbstractType<Short> {
     @Override
     public Short getValue(ResultSet rs, int startIndex) throws SQLException {
         short val = rs.getShort(startIndex);
-        return rs.wasNull() ? null : Short.valueOf(val);
+        return rs.wasNull() ? null : val;
     }
 
     @Override
